Delete Procfile
A Procfile is no longer required to run a Node.js app on Heroku. If no Procfile is present in the root directory of your app during the build process, heroku will check for a scripts.start entry in your package.json file. If such an entry is present, a default Procfile is generated automatically:

$ cat Procfile
web: npm start